move ErrorResponse<T> into HttpPipelineExtensions
