+++ sdk/core/Azure.Core/src/Shared/HttpPipelineExtensions.cs
@@ -90,5 +90,21 @@ namespace Azure.Core
 
             return (requestContext.CancellationToken, requestContext.ErrorOptions);
         }
+
+        internal class ErrorResponse<T> : Response<T>
+        {
+            private readonly Response _response;
+            private readonly RequestFailedException _exception;
+
+            public ErrorResponse(Response response, RequestFailedException exception)
+            {
+                _response = response;
+                _exception = exception;
+            }
+
+            public override T Value { get => throw _exception; }
+
+            public override Response GetRawResponse() => _response;
+        }
     }
 }
