+++ sdk/core/Azure.Core/tests/Azure.Core.Tests.csproj
@@ -35,7 +35,6 @@
     <Compile Include="..\src\Shared\ChangeTrackingDictionary.cs" LinkBase="Shared" />
     <Compile Include="..\src\Shared\ChangeTrackingList.cs" LinkBase="Shared" />
     <Compile Include="..\src\Shared\ConnectionString.cs" LinkBase="Shared" />
-    <Compile Include="..\src\Shared\ErrorResponse.cs" LinkBase="Shared" />
     <Compile Include="..\src\Shared\FormUrlEncodedContent.cs" LinkBase="Shared" />
     <Compile Include="..\src\Shared\ForwardsClientCallsAttribute.cs" LinkBase="Shared" />
     <Compile Include="..\src\Shared\HttpPipelineExtensions.cs" LinkBase="Shared" />
